+++ src/components/recorder/RecorderPanel.jsx
@@ -83,15 +83,22 @@ export default function RecorderPanel({ children, onChange }) {
     onStatus,
   });
 
+  const { inputRef, startCapturing } = useInputCapture();
+
   useEffect(() => {
     if (recorder.active && recorder.forceStart) {
       setRecorderActive(true, false);
-      startRecording();
+      if (isInputCaptureSupported() || !isMediaRecorderSupported()) {
+        startCapturing();
+      } else {
+        startRecording();
+      }
     }
   }, [recorder, setRecorderActive, startRecording]);
 
   return (
     <div className="relative flex flex-col items-center bg-black w-full h-full">
+      <VideoInputCapture inputRef={inputRef} onChange={() => {}} />
       <div className="flex-1" />
       <video className="w-full h-full bg-neutral-900" ref={videoRef} />
       <div className="flex-1" />
@@ -119,16 +126,62 @@ export default function RecorderPanel({ children, onChange }) {
   );
 }
 
-/**
- * @param {string} value
- */
-function isStartingOrRecording(value) {
-  return value === 'starting' || value === 'recording';
+function useInputCapture() {
+  const inputRef = useRef(/** @type {HTMLInputElement|null} */ (null));
+  const startCapturing = useCallback(
+    function startCapturing() {
+      inputRef.current?.click();
+    },
+    [inputRef],
+  );
+  return {
+    inputRef,
+    startCapturing,
+  };
+}
+
+function isInputCaptureSupported() {
+  return (
+    typeof document !== 'undefined' &&
+    typeof document.createElement('input').capture !== 'undefined'
+  );
+}
+
+function isMediaRecorderSupported() {
+  return (
+    typeof window !== 'undefined' && typeof window.MediaRecorder !== 'undefined'
+  );
 }
 
 /**
- * @param {string} value
+ * @param {object} props
+ * @param {import('react').RefObject<HTMLInputElement>} props.inputRef
+ * @param {Function} props.onChange
  */
-function isStopped(value) {
-  return value === 'stopped';
+function VideoInputCapture({ inputRef, onChange }) {
+  const onChangeImpl = useCallback(
+    /** @type {import('react').ChangeEventHandler<HTMLInputElement>} */
+    function _onChangeImpl(e) {
+      const el = /** @type {HTMLInputElement} */ (e.target);
+      const file = el.files?.[0];
+      if (!file) {
+        return;
+      }
+      const url = URL.createObjectURL(file);
+      el.value = '';
+      onChange({ status: 'stopped', data: url });
+    },
+    [onChange],
+  );
+
+  return (
+    <input
+      ref={inputRef}
+      className="hidden"
+      type="file"
+      accept="video/*"
+      capture="environment"
+      onChange={onChangeImpl}
+    />
+  );
 }