+++ src/components/workspace/AppTitle.jsx
@@ -1,0 +1,14 @@
+export default function AppTitle() {
+  return (
+    <h1
+      className={
+        'text-center mx-auto mt-auto underline pointer-events-none whitespace-nowrap' +
+        ' ' +
+        'text-6xl md:text-8xl lg:text-9xl' +
+        ' ' +
+        'mb-4 lg:mb-20'
+      }>
+      [🗒️🎬🎥]
+    </h1>
+  );
+}